requirements.txt
# Core Data Processing
pyspark==3.5.0
delta-spark==3.0.0

# Data Quality
great-expectations==0.18.8

# Geospatial
apache-sedona==1.5.1
geopandas==0.14.2
shapely==2.0.2
pyproj==3.6.1
fiona==1.9.5

# ML
lightgbm==4.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-spark==0.6.0

# Optional: Airflow (commented for minimal setup)
# apache-airflow==2.8.0

# Jupyter (for notebooks)
jupyter==1.0.0
ipykernel==6.27.1

# Data handling
openpyxl==3.1.2
